fix: give news its own entrypoint instead of duplicating repomod load Same behaviour and UX as before, much less tangled loading code:
News no longer runs its own ModRepos.LoadConfig + ListAll early in a
dedicated StageNews (the "repomod entypoint").

- Add NewsRunner.GetActiveNotices(ModList) as the clean data entrypoint.
- News check now happens after the single normal mod list build in
  StageSearching (on firstLoad only).
- Remove cachedNews* fields, InvalidateCachedSearchData, and the
  duplicate loading + cache-stealing hack if the initial implementation
- Migration actions still trigger an immediate list refresh so the
  new repo mod appears before the countdown.
- Updated docs.
