+++ StationeersLaunchPad/LaunchPadConfig.cs
@@ -59,15 +59,6 @@ public static class LaunchPadConfig
 
   private static ModList modList = ModList.NewEmpty();
 
-  private static ModList cachedNewsModList;
-  private static ModReposConfig cachedNewsModRepos;
-
-  public static void InvalidateCachedSearchData()
-  {
-    cachedNewsModList = null;
-    cachedNewsModRepos = null;
-  }
-
   private static LoadStage Stage = LoadStage.Initializing;
   public static bool ModsLoaded => Stage > LoadStage.Configuring;
   public static bool GameRunning => Stage == LoadStage.Running;
@@ -127,7 +118,6 @@ public static class LaunchPadConfig
 
     await StageInitializing();
     await StageUpdating();
-    await StageNews();
 
     var firstLoad = true;
     do
@@ -229,65 +219,6 @@ public static class LaunchPadConfig
       StopAutoLoad();
   }
 
-  private static async UniTask StageNews()
-  {
-    if (Stage == LoadStage.Failed) return;
-    if (Platform.IsServer || !Configs.NewsCheckOnStart.Value) return;
-
-    var url = Configs.EffectiveNewsFeedUrl;
-    if (string.IsNullOrWhiteSpace(url)) return;
-
-    Stage = LoadStage.News;
-    try
-    {
-      Logger.Global.LogInfo("Loading Mod Repos (news)");
-      var modRepos = ModRepos.Current = await ModRepos.LoadConfig();
-
-      Logger.Global.LogInfo("Listing Mods (news)");
-      var defs = await ModSource.ListAll(new()
-      {
-        Repos = modRepos,
-        SteamDisabled = SteamDisabled,
-      });
-      var ml = ModList.FromDefs(defs);
-
-      Logger.Global.LogInfo("Loading Mod Config (news)");
-      var modConfig = ModConfigUtil.LoadConfig();
-      ModRepos.UpdateModPaths(modRepos, modConfig);
-      ml.ApplyConfig(modConfig);
-
-      cachedNewsModRepos = modRepos;
-      cachedNewsModList = ml;
-      modList = ml;
-
-      var feed = await NewsFetcher.Fetch(url);
-      if (feed == null)
-      {
-        Logger.Global.LogDebug("News: no feed or fetch failed");
-        return;
-      }
-
-      Logger.Global.LogDebug($"News: feed has {feed.Entries?.Count ?? 0} entries");
-
-      var dismissed = NewsDismissal.LoadDismissed();
-      var matches = await NewsMatcher.Match(feed, ml, dismissed);
-      if (matches.Count > 0)
-      {
-        Logger.Global.LogInfo($"Displaying {matches.Count} notice(s)");
-        await NewsPopup.Run(matches);
-      }
-      else
-      {
-        Logger.Global.LogDebug("News: no notices matched current enabled mods / dismissed / already-migrated");
-      }
-    }
-    catch (Exception ex)
-    {
-      Logger.Global.LogError("Error in news notices stage (startup will continue)");
-      Logger.Global.LogException(ex);
-    }
-  }
-
   private static async UniTask StageSearching(bool firstLoad)
   {
     if (Stage == LoadStage.Failed) return;
@@ -295,42 +226,39 @@ public static class LaunchPadConfig
     try
     {
       Logger.Global.LogInfo("Loading Mod Repos");
-      ModReposConfig modRepos;
-      if (cachedNewsModRepos != null && cachedNewsModList != null)
-      {
-        modRepos = ModRepos.Current = cachedNewsModRepos;
-        modList = cachedNewsModList;
-        cachedNewsModRepos = null;
-        cachedNewsModList = null;
-        Logger.Global.LogDebug("Using cached mod list from news stage");
-      }
-      else
-      {
-        modRepos = ModRepos.Current = await ModRepos.LoadConfig();
-        if (firstLoad && Configs.RepoCheckUpdates.Value)
-          await ModRepos.UpdateRepos(modRepos);
-        ModRepos.SaveConfig(modRepos);
-        if (firstLoad && Configs.RepoModCheckUpdates.Value)
-        {
-          var updates = ModRepos.GetModUpdateTargets(modRepos);
-          // TODO: prompt user when auto-update not enabled
-          if (updates.Count > 0)
-            await ModRepos.UpdateMods(modRepos, updates);
-          ModRepos.SaveConfig(modRepos);
-        }
+      var modRepos = ModRepos.Current = await ModRepos.LoadConfig();
 
-        Logger.Global.LogInfo("Listing Mods");
-        modList = ModList.FromDefs(await ModSource.ListAll(new()
-        {
-          Repos = modRepos,
-          SteamDisabled = SteamDisabled,
-        }));
+      if (firstLoad && Configs.RepoCheckUpdates.Value)
+        await ModRepos.UpdateRepos(modRepos);
+      ModRepos.SaveConfig(modRepos);
+
+      if (firstLoad && Configs.RepoModCheckUpdates.Value)
+      {
+        var updates = ModRepos.GetModUpdateTargets(modRepos);
+        // TODO: prompt user when auto-update not enabled
+        if (updates.Count > 0)
+          await ModRepos.UpdateMods(modRepos, updates);
+        ModRepos.SaveConfig(modRepos);
       }
+
+      Logger.Global.LogInfo("Listing Mods");
+      var defs = await ModSource.ListAll(new()
+      {
+        Repos = modRepos,
+        SteamDisabled = SteamDisabled,
+      });
+      modList = ModList.FromDefs(defs);
 
       Logger.Global.LogInfo("Loading Mod Config");
       var modConfig = ModConfigUtil.LoadConfig();
       ModRepos.UpdateModPaths(modRepos, modConfig);
       modList.ApplyConfig(modConfig);
+
+      // News check now consumes the single mod list built above.
+      // This removes the duplicate "repomod load" that used to live in StageNews.
+      if (firstLoad)
+        await CheckNewsNotices();
+
       ModConfigUtil.SaveConfig(modList.ToModConfig());
 
       var depNotice = !modList.CheckDependencies();
@@ -346,6 +274,54 @@ public static class LaunchPadConfig
     catch (Exception ex)
     {
       OnStartupError(ex);
+    }
+  }
+
+  private static async UniTask CheckNewsNotices()
+  {
+    if (Platform.IsServer || !Configs.NewsCheckOnStart.Value) return;
+
+    var prevStage = Stage;
+    try
+    {
+      Stage = LoadStage.News;
+
+      var matches = await NewsRunner.GetActiveNotices(modList);
+      if (matches.Count > 0)
+      {
+        Logger.Global.LogInfo($"Displaying {matches.Count} notice(s)");
+        await NewsPopup.Run(matches);
+
+        // A repo_mod_install migration may have added a new tracked repo mod.
+        // Rebuild the snapshot so the new mod appears enabled before the countdown.
+        var repos = ModRepos.Current ?? await ModRepos.LoadConfig();
+        ModRepos.Current = repos;
+
+        var defs = await ModSource.ListAll(new()
+        {
+          Repos = repos,
+          SteamDisabled = SteamDisabled,
+        });
+        modList = ModList.FromDefs(defs);
+
+        var modConfig = ModConfigUtil.LoadConfig();
+        ModRepos.UpdateModPaths(repos, modConfig);
+        modList.ApplyConfig(modConfig);
+      }
+      else
+      {
+        Logger.Global.LogDebug("News: no notices matched current enabled mods / dismissed / already-migrated");
+      }
+    }
+    catch (Exception ex)
+    {
+      Logger.Global.LogError("Error in news notices stage (startup will continue)");
+      Logger.Global.LogException(ex);
+    }
+    finally
+    {
+      if (Stage == LoadStage.News)
+        Stage = prevStage;
     }
   }
